Commit way too big, sorry.  Set out to add support/cases for explicit node type selection to Rails integration and Cucumber tests.  (Can you tell I didn't anticipate node typing as a prerequisite for creating eges, or that I really didn't have any idea what node typing would entail when I started?  Completely backwards WRT (A)TDD....)
In the process of debugging failures in these tests, I figured out what
had been causing the intermittent failures when I tried to do

  HASH[some_node_child.type].new

The problem wasn't in the hash lookup, but in the fact that sometimes
.type evaluated to the ActiveRecord attribute getter method, and
SOMETIMES it was grabbed by Ruby and sent to Object.type (the deprecated
alias of Object.class).  And I'm pretty sure that it was intermittent
based on what ActiveRecord attributes had been called against that model
class previously in the execution order, which is why the problem seemed
intermittent and why all my unit tests passed but it broke when being
served across the web.

So, I did a global rename of the database column that is used to hold
the Node child classes' names for single-table inheritance from "type"
(the Rails default that got me into trouble) to "sti_type".  This
required a new database migration and touching most of the node-model
related files.  Hence the size of this commit.  And, unfortunately,
since I'd already started making test changes and fixing other bugs
before realizing I needed this global change, the deltas are all
intermingled--ick.

We now return you to your regularly-scheduled implementation of Edge MVC
components.....