Adding new columns to allow Connections with scalar (non-Item) objects.
:kind_of_obj is intended as an enumeration to determine what kind of object
(Item, scalar, who-knows-what-in-future) the current Connection has.
Range of enumerated values will be enforced by a model validation (not
done yet), and existing database rows are handled by an UPDATE in the
migration.

The :scalar_obj column will hold the (string) value of the "scalar" that
kind_of_obj=="scalar" Connections "connect" to.  Yes, it's inefficient
to have a string field for every connection just to support those
with scalar objects, but it's "The Simplest Thing That Could Possibly Work"
so I'll deal with some kind of polymorphic model supported by two
different tables later when/if database size becomes a problem.